Move from ontohub_models to ontohub-models.
