+++ lib/ontohub-models.rb
@@ -1,3 +1,7 @@
 # frozen_string_literal: true
 
-require_relative 'ontohub_models'
+require 'ontohub-models/engine'
+
+# Namespace module
+module OntohubModels
+end
