--- lib/ontohub_models.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-require 'ontohub_models/engine'
-
-# Namespace module
-module OntohubModels
-end
