+++ spec/support/simplecov.rb
@@ -12,7 +12,7 @@ if RUBY_ENGINE == 'ruby' # not 'rbx'
       # This initializer is not executed because no migrations are run.
       add_filter 'config/initializers/expose_migrations_path.rb'
       # The version is only used in the gemspec.
-      add_filter 'lib/ontohub_models/version.rb'
+      add_filter 'lib/ontohub-models/version.rb'
       # Ignore the dummy application that is used for the spec.
       add_filter 'spec/dummy'
       # Ignore the RSpec configuration files.
